Implemented routers.
