+++ lib/app_creator.js
@@ -10,9 +10,9 @@ var sockjs = require('sockjs');
 var locale = require('locale');
 
 var utils = require('./utils.js');
-var ipsParser = require('./ips_parser.js');
+var httpRouter = require('./http_router.js');
 var sockHandler = require('./sock_handler.js');
-var routeManager = require('./routes/route_manager.js');
+var routeManager = require('./routers/route_manager.js');
 var appconfigDefault = require('./default/appconfig.js');
 
 module.exports = function(appRouter, path){
@@ -166,7 +166,8 @@ module.exports = function(appRouter, path){
 	// create routes
 	var routers = routeManager(app);
 
-	// TODO serve order
+	// http req router
+	httpRouter(app);
 
 	// API
 	app.setConfig = function(config){
@@ -180,6 +181,7 @@ module.exports = function(appRouter, path){
 		for(var i=0; i<app.config.app.host.length; i++) {
 			appRouter.bindHost(app, app.config.app.host[i]);
 		}
+		// TODO generate client code cache
 		initDb(cb);
 	};
 	app.stop = function(cb){
