+++ lib/default/appconfig.js
@@ -11,6 +11,7 @@ module.exports = {
 	},
 
 	client: {
+		cache: '', // the client code cache dir
 		favicon: '', // the favicon's path in server
 		loadingLogo: '', // the loading logo's path in server
 		loadingLogoBackground: '#fff', // the loading logo's background color
