+++ lib/http_router.js
@@ -2,447 +2,22 @@
 'use strict';
 
 var fs = require('fs');
-var path = require('path');
-var mkdirp = require('mkdirp');
-var uglifyjs = require('uglify-js');
-var handlebars = require('handlebars');
-var fsListener = require('./fs_listener.js');
-var preprocessor = require('./preprocessor.js');
+var express = require('express');
 
-// rmdirp
-var rmdirp = function(path) {
-	var files = [];
-	if(fs.existsSync(path)) {
-		files = fs.readdirSync(path);
-		files.forEach(function(file,index){
-			var curPath = path + '/' + file;
-			if(fs.statSync(curPath).isDirectory()) {
-				rmdirp(curPath);
-			} else {
-				fs.unlinkSync(curPath);
-			}
-		});
-		fs.rmdirSync(path);
-	}
-};
-
-// preprocess client files
-var clientFiles = function(app, cb){
-	if(app.debug) {
-		// run with client code in memory
-		var clientFiles = app.clientFileCache = {};
-		var clientFileHandler = function(file, type, match){
-			if(app.config.app.script === app.config.app.routes && path.basename(file) === 'routes.js') return;
-			if(match === '.locale/') {
-				if(!type) return;
-				file = file.slice(0, file.lastIndexOf('.locale')) + '.tmpl';
-			}
-			var cache = file.split(path.sep).slice(1).join('/');
-			if(type !== 'removed') {
-				preprocessor(app, file, function(res){
-					if(!res) return;
-					for(var k in res)
-						clientFiles[cache+k] = res[k];
-				});
-				if(type) console.log('Client Code Updated: '+file);
-			} else {
-				delete clientFiles[cache];
-			}
-		};
-		var clientStaticHandler = function(){};
-	} else if(app.mode === 'CACHE') {
-		// run with cache
-		rmdirp(app.config.path.cache + '/client/');
-		var clientFileHandler = function(file, type, match){
-			if(type) return;
-			if(app.config.app.script === app.config.app.routes && path.basename(file) === 'routes.js') return;
-			if(match === '.locale/') return;
-			// find .min
-			var cache = app.config.path.cache + '/client/' + file.split(path.sep).slice(1).join('/');
-			mkdirp.sync(path.dirname(cache));
-			if(file.slice(-7) === '.min.js') {
-				if(fs.existsSync(file.slice(-7) + '.js')) return;
-				fs.writeFileSync(cache, fs.readFileSync(file));
-				return;
-			}
-			if(file.slice(-8) === '.min.css') {
-				if(fs.existsSync(file.slice(-8) + '.css')) return;
-				fs.writeFileSync(cache, fs.readFileSync(file));
-				return;
-			}
-			var minFile = '';
-			if(path.extname(file) === '.js')
-				minFile = file.slice(0, -3) + '.min.js';
-			else if(path.extname(file) === '.css')
-				minFile = file.slice(0, -4) + '.min.css';
-			// write
-			if(minFile && fs.existsSync(minFile)) {
-				// use .min instead of compressing
-				fs.writeFileSync(cache, fs.readFileSync(minFile));
-			} else {
-				// compress
-				preprocessor(app, file, function(res){
-					if(!res) return;
-					for(var k in res)
-						fs.writeFileSync(cache+k, res[k]);
-				});
-			}
-		};
-		var clientStaticHandler = function(file, type, match){
-			if(type) return;
-			if(file.match(/\.(js|tmpl|css|stylus|locale[\/\\].+)$/)) return;
-			var cache = app.config.path.cache + '/client/' + file.split(path.sep).slice(1).join('/');
-			mkdirp.sync(path.dirname(cache));
-			fs.writeFileSync(cache, fs.readFileSync(file));
-		};
-	} else {
-		// run with previous cached code
-		var clientFileHandler = null;
-	}
-	if(clientFileHandler) {
-		var c = 4;
-		var finished = function(){
-			c--;
-			if(!c) cb();
-			return !!app.debug;
-		};
-		// listen for changes in debug or cache mode
-		fsListener(app.config.path.client, {
-			'*': clientStaticHandler,
-		}, finished);
-		fsListener(app.config.path.script, {
-			'.js': clientFileHandler,
-		}, finished);
-		fsListener(app.config.path.tmpl, {
-			'.tmpl': clientFileHandler,
-			'.locale/': clientFileHandler,
-		}, finished);
-		fsListener(app.config.path.style, {
-			'.css': clientFileHandler,
-			'.stylus': clientFileHandler
-		}, finished);
-	} else {
-		cb();
-	}
-};
-
-// generate client routes
-var clientRoutes = function(app, cb){
-	var routes = app.clientRoutes = {};
-	// listen for changes
-	fsListener(app.config.path.routes, {
-		'routes.js': function(file, type){
-			if(type && !app.debug) return;
-			if(type !== 'removed') {
-				try {
-					delete require.cache[app.cwd+'/'+file];
-					var basepath = file.split(path.sep).slice(1, -1).join('/');
-					if(basepath) basepath += '/';
-					fw.currentLoading = app.config.path.script + '/' + basepath + 'routes.js';
-					routes[basepath] = require(app.cwd+'/'+file);
-					if(app.debug && type) console.log('Routes Updated: '+file);
-				} catch(e) {
-					delete routes[file];
-					console.trace(e);
-				}
-				fw.currentLoading = '';
-			} else
-				delete routes[file];
-			if(type) updateRoutesFile(app);
-		}
-	}, function(){
-		cb();
-		if(!app.debug) return false;
-	});
-};
-
-// generate pages routes
-var pageRoutes = function(app, cb){
-	var routes = app.pageRoutes;
-	fsListener(app.config.path.page, {
-		'.js': function(file, type){
-			if(type && !app.debug) return;
-			var pagePath = '/' + file.split(path.sep).slice(1).join('/').slice(0, -3);
-			if(type !== 'removed') {
-				try {
-					delete require.cache[app.cwd+'/'+file];
-					fw.currentLoading = app.config.path.page + pagePath + '.js';
-					routes[pagePath] = require(app.cwd+'/'+file);
-					if(app.debug && type) console.log('Page Updated: '+pagePath);
-				} catch(e) {
-					delete routes[pagePath];
-					console.trace(e);
-				}
-				fw.currentLoading = '';
-			} else
-				delete routes[pagePath];
-			if(type) updateRoutesFile(app);
-		},
-		'.tmpl': function(file, type){
-			if(!type || !app.debug) return;
-			app.restart();
-		},
-		'.locale/': function(file, type){
-			if(!type || !app.debug) return;
-			app.restart();
-		}
-	}, function(){
-		cb();
-		if(!app.debug) return false;
-	});
-};
-
-// cache routes json string
-var routesFile = '';
-var routeTree = null;
-var routeParser = function(client, page){
-	var tree = {
-		client: client,
-		c: {}
-	};
-	var treeChild = function(node, index){
-		if(node.hasOwnProperty.call(node.c, index))
-			return node.c[index];
-	};
-	var treeAdd = function(path, v){
-		var cur = tree;
-		var segs = path.split('/');
-		while(segs.length) {
-			var seg = segs.shift();
-			if(!seg) continue;
-			if(seg === '*') {
-				cur.d = {};
-				cur = cur.d;
-				break;
-			}
-			if(seg.charAt(0) === ':') seg = ':';
-			var child = treeChild(cur, seg);
-			if(!child)
-				child = cur.c[seg] = {
-					c: {}
-				};
-			cur = child;
-		}
-		cur.v = v;
-	};
-	for(var k in client) {
-		if(k.charAt(0) !== '/') continue;
-		var keys = k.match(/\/(:\w*|\*)/g);
-		if(keys) {
-			for(var i=0; i<keys.length; i++)
-				keys[i] = (keys[i] === '/*' ? '*' : keys[i].slice(2));
-			client[k].k = keys;
-		}
-		treeAdd(k, k);
-	}
-	for(var k in page)
-		treeAdd(k, '');
-	return tree;
-};
-var pathParser = function(tree, path){
-	while(1) {
-		// find node
-		var cur = tree;
-		var def = null;
-		var defSegs = [];
-		var argVals = [];
-		var args = {};
-		var segs = path.split('/');
-		while(segs.length) {
-			if(cur.d) {
-				def = cur.d;
-				defSegs = [];
-			}
-			var seg = segs.shift();
-			if(!seg) continue;
-			defSegs.push(seg);
-			var child = cur.hasOwnProperty.call(cur.c, seg);
-			if(child) {
-				cur = cur.c[seg];
-			} else if(cur.c[':']) {
-				cur = cur.c[':'];
-				argVals.push(seg);
-			} else {
-				cur = def;
-				args['*'] = defSegs.concat(segs).join('/');
-				break;
-			}
-		}
-		if(!cur) return null;
-		if(typeof(cur.v) !== 'string') {
-			if(!def) return null;
-			cur = def;
-			args['*'] = defSegs.concat(segs).join('/');
-		}
-		if(cur.v === '') return 'page';
-		// get obj
-		var id = cur.v;
-		var obj = tree.client[id];
-		if(id.slice(-1) === '/' && path.slice(-1) !== '/')
-			path += '/';
-		else if(id.slice(-1) !== '/' && path.slice(-1) === '/')
-			path = path.slice(0, -1) || '/';
-		// get args
-		if(obj.k)
-			for(var i=0; i<obj.k.length; i++)
-				if(obj.k[i] !== '*') args[obj.k[i]] = argVals[i];
-		// redirect
-		if(!obj.redirect) break;
-		path = obj.redirect.replace(/\/:\w*/g, function(m){
-			return '/'+args[m.slice(2)];
-		});
-	}
-	return {
-		path: path,
-		id: id,
-		route: obj,
-		args: args
-	};
-};
-var pathParserStr = pathParser.toString();
-var updateRoutesFile = function(app){
-	var client = app.clientRoutes;
-	var clientRoutes = {};
-	// combine client routes
-	for(var base in client) {
-		for(var id in client[base]) {
-			var route = client[base][id];
-			if(id.indexOf('/') >= 0 || id.indexOf('*') >= 0) {
-				// relative path routing
-				if(id.charAt(0) !== '/') id = ('/' + base + id).replace(/\/\.\//g, '/');
-			} else {
-				id += ':/' + base;
-			}
-			if(route.redirect) {
-				// redirect
-				var r = route.redirect;
-				if(r.charAt(0) !== '/') route.redirect = ('/' + base + r).replace(/\/\.\//g, '/');
-			}
-			route.base = base;
-			clientRoutes[id] = route;
-			// scan for minify path in file lists
-			if(!app.debug) {
-				var combineFiles = function(files, nameFunc){
-					if(!files) return;
-					if(files.constructor !== Array) files = [files];
-					for(var i=0; i<files.length; i++) {
-						var file = files[i];
-						if(!file.minify) continue;
-						var out = nameFunc(file.minify);
-						if(out.charAt(0) !== '/')
-							out = app.config.path.cache + '/client/' + route.base + out;
-						else
-							out = app.config.path.cache + '/client' + out;
-						mkdirp.sync(path.dirname(out));
-						var fd = fs.openSync(out, 'w');
-						if(file.src.constructor === Array) {
-							for(var j=0; j<files[i].src.length; j++) {
-								if(j) {
-									var buf = new Buffer('\n');
-									fs.writeSync(fd, buf, 0, buf.length);
-								}
-								if(file.src[j].charAt(0) !== '/')
-									buf = fs.readFileSync(app.config.path.cache + '/client/' + route.base + nameFunc(file.src[j]));
-								else
-									buf = fs.readFileSync(app.config.path.cache + '/client' + nameFunc(file.src[j]));
-								fs.writeSync(fd, buf, 0, buf.length);
-							}
-						} else {
-							if(file.src.charAt(0) !== '/')
-								var buf = fs.readFileSync(app.config.path.cache + '/client/' + route.base + nameFunc(file.src));
-							else
-								var buf = fs.readFileSync(app.config.path.cache + '/client' + nameFunc(file.src));
-							fs.writeSync(fd, buf, 0, buf.length);
-						}
-						fs.closeSync(fd);
-					}
-				};
-				combineFiles(route.main, function(file){
-					if(file.slice(-3) !== '.js') file += '.js';
-					return file;
-				});
-				combineFiles(route.lib, function(file){
-					if(file.slice(-3) !== '.js') file += '.js';
-					return file;
-				});
-				combineFiles(route.style, function(file){
-					if(file.slice(-4) !== '.css') {
-						if(file.slice(-11) !== '.stylus.css')
-							file += '.stylus.css';
-						else
-							file += '.css';
-					}
-					return file;
-				});
-				combineFiles(route.tmpl, function(file){
-					if(file.slice(-5) !== '.tmpl')
-						file += '.tmpl';
-					file += '.js';
-					return file;
-				});
-				var locale = app.config.app.locale;
-				for(var i=0; i<locale.length; i++) {
-					var loc = locale[i];
-					combineFiles(route.tmpl, function(file){
-						if(file.slice(-5) !== '.tmpl')
-							file += '.tmpl';
-						if(loc)
-							file += '.' + loc;
-						file += '.js';
-						return file;
-					});
-				}
-			}
-		}
-	}
-	// calculate parent
-	for(var id in clientRoutes) {
-		var route = clientRoutes[id];
-		if(!route.parent) continue;
-		var parent = route.parent;
-		if(parent.indexOf('/') >= 0 || parent.indexOf('*') >= 0) {
-			if(parent.charAt(0) !== '/') id = ('/' + route.base + id).replace(/\/.\//g, '/');
-		} else {
-			var i = ':/' + route.base;
-			while(1) {
-				if(clientRoutes[parent+i]) {
-					route.parent += i;
-					break;
-				}
-				var p = i.lastIndexOf('/', i.length-2) + 1;
-				if(!p) {
-					route.parent = '';
-					if(app.debug) console.error('fw root error: no parent page "'+parent+'" found for "'+id+'".');
-					break;
-				}
-				i = i.slice(0, p);
-			}
-		}
-	}
-	routeTree = routeParser(clientRoutes, app.pageRoutes);
-	routesFile = 'fw._routes(' + JSON.stringify(routeTree) + ', ' + pathParserStr + ')';
-	if(app.mode === 'CACHE')
-		fs.writeFileSync(app.config.path.cache+'/fw/routes.js', uglifyjs.minify(routesFile, {fromString: true}).code);
-	app.pathParser = pathParser;
-	app.routeTree = routeTree;
-};
-
-exports.route = function(app, express, cb){
-	// init routes
-	if(app.debug) console.log('Generating routes...');
-	app.clientRoutes = {};
-	app.pageRoutes = {};
+module.exports = function(app, cb){
+	var expressApp = app.express;
 
 	// default cache
-	if(app.debug)
-		app.use(function(req, res, next){
+	if(fw.debug) {
+		expressApp.use(function(req, res, next){
 			res.set('Cache-Control', 'no cache, no store');
 			next();
 		});
+	}
 
 	// static cache
 	var staticCache = function(req, res, next){
-		if(!app.debug && req.query.v) {
+		if(!fw.debug && req.query.v) {
 			if(req.query.v !== app.config.app.version) {
 				res.send(481, 'Version Not Match');
 				return;
@@ -452,64 +27,58 @@ exports.route = function(app, express, cb){
 		}
 		next();
 	};
-	app.use('/~fw', staticCache);
-	app.use('/~', staticCache);
+	expressApp.use('/~fw', staticCache);
+	expressApp.use('/~client', staticCache);
 
 	// prevent cache
 	var preventCache = function(req, res, next){
 		res.set('Cache-Control', 'no-cache, no-store');
 		next();
 	};
-	app.use('/~conf', preventCache);
-	app.use('/~render', preventCache);
+	expressApp.use('/~conf', preventCache);
+	expressApp.use('/~render', preventCache);
 
-	// special files
-	app.use('/', express.static(app.config.path.static));
+	// fw files
+	if(fw.debug) {
+		expressApp.use('/~fw', express.static(__dirname + '/client'));
+	} else {
+		expressApp.use('/~fw', express.static(app.config.client.cache + '/~fw'))
+	}
 
 	// client files
-	if(app.debug) {
-		app.use('/~fw', express.static(__dirname+'/client'));
-		app.use('/~', function(req, res, next){
-			var file = req.path.slice(1);
-			if(typeof(app.clientFileCache[file]) === 'undefined') {
-				next();
+	if(fw.debug) {
+		expressApp.use('/~client', function(req, res, next){
+			// TODO check req path and file ext
+			if(req.method !== 'GET' && req.method !== 'HEAD') {
+				res(404);
 				return;
 			}
-			var ext = path.extname(file);
-			if(ext === '.js') res.type('text/javascript');
-			else if(ext === '.css') res.type('text/css');
-			res.send(app.clientFileCache[file]);
+			var reqPath = req.path;
+			var file = app.route.client(reqPath);
+			if(!file) {
+				res(404);
+				return;
+			}
+			if(file.match(/\.(js|tmpl|css|stylus)$/)) {
+				// direct send
+				res.sendFile(file);
+				return;
+			}
+			preprocessor(app, file, function(files){
+				if(file.slice(-5) === '.tmpl') {
+					// match lang for tmpl files
+					res(files['.'+req.language+'.js'] || files['.js']);
+				} else {
+					// get the only file
+					for(var k in files) {
+						res(files[k]);
+						break;
+					}
+				}
+			});
 		});
-		app.use('/~', express.static(app.config.path.script));
 	} else {
-		if(app.mode === 'CACHE') {
-			// write cache/fw
-			rmdirp(app.config.path.cache + '/fw');
-			if(!fs.existsSync(app.config.path.cache))
-				fs.mkdir(app.config.path.cache);
-			if(!fs.existsSync(app.config.path.cache + '/fw'))
-				fs.mkdirSync(app.config.path.cache + '/fw');
-			if(app.config.client.loadingLogo)
-				fs.writeFileSync(app.config.path.cache+'/fw/'+app.config.client.loadingLogo, fs.readFileSync(app.config.path.rc+'/'+app.config.client.loadingLogo));
-			if(app.config.client.favicon)
-				fs.writeFileSync(app.config.path.cache+'/fw/'+app.config.client.favicon, fs.readFileSync(app.config.path.rc+'/'+app.config.client.favicon));
-			// compress fw code
-			var parseCopyright = function(str){
-				var copyright = '';
-				var m = str.match(/^\s*(\/\*.+?\*\/)/);
-				if(m) copyright = m[1] + '\n';
-				m = str.match(/^\s*(\/\/.+?)[\r\n]/);
-				if(m) copyright = m[1] + '\n';
-				return copyright;
-			};
-			var str = fs.readFileSync(__dirname+'/client/init.js').toString('utf8');
-			fs.writeFileSync(app.config.path.cache+'/fw/init.js', parseCopyright(str) + uglifyjs.minify(str, {fromString: true}).code);
-			fs.writeFileSync(app.config.path.cache+'/fw/libs.js', fs.readFileSync(__dirname+'/client/libs.js'));
-			var str = fs.readFileSync(__dirname+'/client/fw.js').toString('utf8');
-			fs.writeFileSync(app.config.path.cache+'/fw/fw.js', parseCopyright(str) + uglifyjs.minify(str, {fromString: true}).code);
-		}
-		app.use('/~fw', express.static(app.config.path.cache + '/fw'));
-		app.use('/~', express.static(app.config.path.cache + '/client'));
+		expressApp.use('/~client', express.static(app.config.client.cache + '/~client'));
 	}
 
 	// index page generator for web
@@ -534,10 +103,18 @@ exports.route = function(app, express, cb){
 			auth: auth,
 		});
 	};
+	app.sendRender = function(req, res){
+		// default file
+		res.type('text/html');
+		res.set('Cache-Control', 'no-cache, no-store');
+		app.render.path(app, req, req.path, 0, function(css, r){
+			res.status(r.statusCode).send(indexPage(app.sessionCookieVal(req.sessionID), req.language, css, r.extraHead, r.title, r.content));
+		});
+	};
 
 	// index page generator for app
-	if(app.mode === 'CACHE') {
-		var packaged = handlebars.compile(fs.readFileSync( __dirname+'/default/packaged_webapp.html' ).toString('utf8'));
+	if(fw.mode === 'cache') {
+		var packaged = handlebars.compile(fs.readFileSync( __dirname+'/default/webapp.html' ).toString('utf8'));
 		var html = packaged({
 			debug: app.debug || 0,
 			version: app.config.app.version,
@@ -551,123 +128,98 @@ exports.route = function(app, express, cb){
 			title: app.config.app.title,
 			meta: app.config.client.meta,
 		});
-		fs.writeFileSync(app.config.path.cache+'/webapp.html', html);
+		fs.writeFileSync(app.config.client.cache+'/webapp.html', html);
 	}
 
 	// default routes
-	app.get('*', function(req, res){
-		if(app.debug && req.path === '/~fw/routes.js') {
-			// routes file
-			res.type('text/javascript');
-			res.send(routesFile);
-		} else if(app.debug && req.path === '/~fw/'+app.config.client.loadingLogo) {
-			// loading logo
-			res.sendFile(app.config.path.rc + '/' + app.config.client.loadingLogo, {root: process.cwd()});
-		} else if(app.debug && req.path === '/~fw/'+app.config.client.favicon) {
-			// loading logo
-			res.sendFile(app.config.path.rc + '/' + app.config.client.favicon, {root: process.cwd()});
-		} else if(req.path === '/~conf/sock.js') {
-			// sock config file
-			res.type('text/javascript');
-			if(!app.debug && req.query.v && req.query.v !== app.config.app.version) {
-				// server version changed
-				res.send('fw.onserverchanged("'+app.config.app.version+'")');
-				return;
-			}
-			if(!req.query.a) {
-				// generate new session
-				app.generateSession(function(id){
-					res.send('fw._sockConfig("' + id + '")');
-				});
-				return;
-			}
-			app.getAuthSession(req.query.a, function(session){
-				if(!session) {
-					// session lost
-					res.send('fw._sessionlost()');
+	expressApp.get('*', function(req, res){
+		if(req.path.slice(0, 2) === '/~') {
+			if(fw.debug && req.path === '/~fw/routes.js') {
+				// routes file
+				res.type('text/javascript');
+				// TODO generate client routes file
+				res.send(app.clientRoutesFile);
+			} else if(fw.debug && req.path === '/~fw/'+app.config.client.loadingLogo) {
+				// loading logo
+				res.sendFile(app.config.client.loadingLogo, {root: process.cwd()});
+			} else if(fw.debug && req.path === '/~fw/'+app.config.client.favicon) {
+				// loading logo
+				res.sendFile(app.config.client.favicon, {root: process.cwd()});
+			} else if(req.path === '/~conf/sock.js') {
+				// sock config file
+				res.type('text/javascript');
+				if(!fw.debug && req.query.v && req.query.v !== app.config.app.version) {
+					// server version changed
+					res.send('fw.onserverchanged("'+app.config.app.version+'")');
 					return;
 				}
-				res.send('fw._sockConfig()');
-			});
-		} else if(req.path === '/~conf/lang') {
-			// set language, must in same-origin
-			var auth = req.query.a || '';
-			if(req.query.a !== req.cookies['fw.sid']) {
-				res.send(403);
-				return;
-			}
-			var lang = req.query.l;
-			if(typeof(lang) === 'string')
-				req.selectLanguage(lang);
-			else
-				req.selectLanguage();
-			if(req.query.r) {
-				res.redirect(req.query.r || '/');
-			} else {
-				res.type('text/javascript');
-				res.send(';');
-			}
-		} else if(req.path.slice(0, 8) === '/~render') {
-			// rendering request
-			var path = req.path.slice(8);
-			var depth = req.query.d;
-			if(!path || !depth) {
-				res.send(403);
-				return;
-			}
-			if(typeof(req.query.l) === 'string')
-				req.language = req.conn.language = req.query.l;
-			app.getAuthSession(req.query.a, function(session){
-				if(!session) {
-					// session lost
+				if(!req.query.a) {
+					// generate new session
+					app.generateSession(function(id){
+						res.send('fw._sockConfig("' + id + '")');
+					});
+					return;
+				}
+				app.getAuthSession(req.query.a, function(session){
+					if(!session) {
+						// session lost
+						res.send('fw._sessionlost()');
+						return;
+					}
+					res.send('fw._sockConfig()');
+				});
+			} else if(req.path === '/~conf/lang') {
+				// set language, must in same-origin
+				var auth = req.query.a || '';
+				if(req.query.a !== req.cookies['fw.sid']) {
 					res.send(403);
 					return;
 				}
-				req.sessionID = req.query.a;
-				req.session = req.conn.session = session;
-				res.type('text/javascript');
-				app.render.path(app, req, path, depth, function(css, r){
-					res.send('fw._renderRes=' + JSON.stringify(r));
+				var lang = req.query.l;
+				if(typeof(lang) === 'string')
+					req.selectLanguage(lang);
+				else
+					req.selectLanguage();
+				if(req.query.r) {
+					res.redirect(req.query.r || '/');
+				} else {
+					res.type('text/javascript');
+					res.send(';');
+				}
+			} else if(req.path.slice(0, 9) === '/~render/') {
+				// rendering request
+				var path = req.path.slice(8);
+				var depth = req.query.d;
+				if(!path || !depth) {
+					res.send(403);
+					return;
+				}
+				if(typeof(req.query.l) === 'string')
+					req.language = req.conn.language = req.query.l;
+				app.getAuthSession(req.query.a, function(session){
+					if(!session) {
+						// session lost
+						res.send(403);
+						return;
+					}
+					req.sessionID = req.query.a;
+					req.session = req.conn.session = session;
+					res.type('text/javascript');
+					app.render.path(app, req, path, depth, function(css, r){
+						res.send('fw._renderRes=' + JSON.stringify(r));
+					});
 				});
-			});
-		} else if(req.path.slice(0, 2) === '/~') {
-			// prevent protected path
-			res.send(404, '');
-		} else if(app.pageRoutes[req.path]) {
-			// special page
-			app.pageRoutes[req.path](req, res);
+			} else {
+				// prevent protected path
+				res.send(404);
+			}
 		} else {
-			// default file
-			res.type('text/html');
-			res.set('Cache-Control', 'no-cache, no-store');
-			app.render.path(app, req, req.path, 0, function(css, r){
-				res.status(r.statusCode).send(indexPage(app.sessionCookieVal(req.sessionID), req.language, css, r.extraHead, r.title, r.content));
-			});
+			app.route.static(req.path);
 		}
 	});
 
 	// head, post, put, delete, options, trace for special pages
 	app.all('*', function(req, res){
-		if(app.pageRoutes[req.path]) {
-			// special page
-			app.pageRoutes[req.path](req, res);
-		} else {
-			res.send(404, '');
-		}
+		app.route.static(req.path);
 	});
-
-	// dynamic contents
-	var c = 4;
-	var finished = function(){
-		c--;
-		if(!c) {
-			updateRoutesFile(app);
-			app.render = require('./render.js');
-			app.render.init(app, cb);
-		}
-	};
-	clientFiles(app, finished);
-	clientRoutes(app, finished);
-	pageRoutes(app, finished);
-	finished();
 };
