+++ lib/main.js
@@ -1,11 +1,9 @@
 // Copyright 2014 LastLeaf, LICENSE: github.lastleaf.me/MIT
 'use strict';
 
-var fs = require('fs');
 var http = require('http');
 
 var appCreator = require('./app_creator.js');
-var tmplParser = require('./tmpl_parser.js').parser;
 
 // create global fw object
 var fw = global.fw = {};
@@ -23,8 +21,8 @@ var fwObjInit = function(){
 	fw.currentLoading = null;
 	fw.tmpl = function(file){
 		if(!fw.currentLoading) return null;
-		file = fw.currentLoading.slice(0, fw.currentLoading.lastIndexOf('/')+1) + file;
-		return tmplParser(app, fs.readFileSync(file).toString('utf8'), file);
+		file = fw.currentLoading.path.slice(0, fw.currentLoading.lastIndexOf('/')+1) + file;
+		return fw.currentLoading.app.route.tmpl.generate(fw.currentLoading.type, file);
 	};
 	fw.module = function(name){
 		// TODO module path
