+++ lib/preprocessor.js
@@ -9,7 +9,7 @@ var uglifyjs = require('uglify-js');
 var tmplPreparser = require('./tmpl_parser.js').preparser;
 
 module.exports = function(app, file, cb){
-	var res = null;
+	var res = {};
 	var ext = path.extname(file).slice(1);
 	try {
 		var str = fs.readFileSync(file).toString('utf8');
@@ -31,7 +31,7 @@ var handlers = {};
 handlers.tmpl = function(app, str, file, cb){
 	var obj = tmplPreparser(app, str, file);
 	var copyright = '';
-	if(!app.debug) {
+	if(!fw.debug) {
 		var m = str.match(/^\s*\<!--([\s\S]+?)--\>/);
 		if(m) copyright = '/*' + m[1] + '*/\n';
 	}
@@ -51,7 +51,7 @@ handlers.stylus = function(app, str, file, cb){
 
 // css
 handlers.css = function(app, str, file, cb){
-	if(!app.debug) {
+	if(!fw.debug) {
 		var copyright = '';
 		var m = str.match(/^\s*(\/\*[\s\S]+?\*\/)/);
 		if(m) copyright = m[1] + '\n';
@@ -65,7 +65,7 @@ handlers.css = function(app, str, file, cb){
 
 // js
 handlers.js = function(app, str, file, cb){
-	if(!app.debug) {
+	if(!fw.debug) {
 		var copyright = '';
 		var m = str.match(/^\s*(\/\*[\s\S]+?\*\/)/);
 		if(m) copyright = m[1] + '\n';
