+++ lib/routers/client.js
@@ -1,4 +1,5 @@
 // Copyright 2014 LastLeaf, LICENSE: github.lastleaf.me/MIT
 'use strict';
 
+// TODO
 module.exports = function(app, prefix, path, cb){};
