+++ lib/routers/helpers/tmpl_parser.js
@@ -99,9 +99,7 @@ exports.parser = function(app, str, file){
 				res[k][tmpl[2]] = handlebars.compile(translated[k]);
 		}
 	}
-	return function(conn){
-		return res[conn.language] || res[''];
-	};
+	return res;
 };
 
 exports.generateLocaleFiles = function(app, str, file, locale, cb){
