+++ lib/routers/render.js
@@ -1,4 +1,8 @@
 // Copyright 2014 LastLeaf, LICENSE: github.lastleaf.me/MIT
 'use strict';
 
-module.exports = function(app, prefix, path, cb){};
+var serverRouter = require('./helpers/server_router.js');
+
+module.exports = function(app, prefix, path, cb){
+	serverRouter(app, 'render', prefix, path, null, false, cb);
+};
