+++ lib/routers/static.js
@@ -3,15 +3,21 @@
 
 var send = require('send');
 
-module.exports = function(app, rootPath, cb){
+module.exports = function(app, prefix, rootPath, cb){
 	setTimeout(function(){
 		cb(null, {
 			dir: rootPath,
 			route: function(next, path, req, res){
 				send(req, path, {index: false, root: rootPath})
 					.on('error', function(err){
-						if(next) next();
-						else res(404);
+						if(next) {
+							next();
+						} else {
+							// route special pages
+							var func = app.route.page(path);
+							if(funcs) funcs.call(app, req, res);
+							else app.sendRender(req, res);
+						}
 					})
 					.pipe(res);
 				return true;
