+++ lib/sock_handler.js
@@ -40,6 +40,7 @@ var connRpc = exports.connRpc = function(app, conn){
 			errCb.apply(cb, a);
 		};
 		cb.next = function(){
+			// TODO support return-back filter
 			var func = funcs.shift();
 			if(func) func.apply(conn, args);
 		};
